// src/types/testType.ts
export interface testType{
    practical_apply_duration: string; 
    practical_result_duration: string; 
    practical_test_start: string; 
    practical_test_end: string; 
    session_for_year: string; 
    test_category: string; 
    test_id: number; 
    written_apply_duration: string; 
    written_result_duration: string; 
    written_test_start: string; 
    written_test_end: string; 
}